完善自动备份功能脚本

+++ src/system/crontab/_ctl/Backupper.class.php
@@ -1,10 +1,19 @@
 <?php
+/**
+ * 自动备份脚本，备份MySQL数据库，通过SSH+Rsync备份文件目录
+ *
+ * @author john
+ */
 namespace Lge;
 
 if (!defined('LGE')) {
 	exit('Include Permission Denied!');
 }
 
+/**
+ * Class Controller_Backupper
+ * @package Lge
+ */
 class Controller_Backupper extends Controller_Base
 {
     public $logCategory = 'crontab/backupper';
@@ -41,12 +50,19 @@ class Controller_Backupper extends Controller_Base
                         $filePathTemp  = "{$dataBackupDir}/{$name}.{$date}.temp.sql.bz2";
                         $localShellCmd = "mysqldump -C -h{$host} -P{$port} -u{$user} -p{$pass} {$name} | bzip2 > {$filePathTemp}";
                         Logger::log("Backing up database: {$filePath}", $this->logCategory);
-                        shell_exec($localShellCmd);
-                        // 判断是否备份成功(大于1K)，使用临时文件防止失败时被覆盖
-                        if (filesize($filePathTemp) > 1024) {
-                            copy($filePathTemp, $filePath);
+
+                        try {
+                            $result = @shell_exec($localShellCmd);
+                            Logger::log($result, $this->logCategory);
+                        } catch (\Exception $e) { }
+
+                        if (file_exists($filePathTemp)) {
+                            // 判断是否备份成功(大于1K)，使用临时文件防止失败时被覆盖
+                            if (filesize($filePathTemp) > 1024) {
+                                copy($filePathTemp, $filePath);
+                            }
+                            unlink($filePathTemp);
                         }
-                        unlink($filePathTemp);
                         if ($keepDays > 1) {
                             $this->_clearDirByKeepDays($dataBackupDir, $keepDays);
                         }
@@ -68,21 +84,29 @@ class Controller_Backupper extends Controller_Base
                         $pass = $clientConfig['pass'];
                         $folderPath = rtrim($folderPath, '/');
                         $folderName = basename($folderPath);
-                        $sshShellCmds = array(
-                            array("rsync -aurvz --delete -e 'ssh -p {$port}' {$folderPath} {$user}@{$host}:{$fileBackupDir}", false, 10),
-                            array("yes", true, 10),
-                            array($pass, true, 3600),
-                        );
                         try {
-                            $ssh = new Lib_Network_Ssh($fileConfig['host'], $fileConfig['port'], $fileConfig['user'], $fileConfig['pass']);
-                            $ssh->asyncCmd($sshShellCmds);
+                            // 先把本地的目录做备份
                             if ($keepDays > 0) {
                                 $backupDirPath = rtrim($fileBackupDir, '/').'/'.$folderName;
                                 if (file_exists($backupDirPath)) {
-                                    $this->_compressBackupFileDir($backupDirPath);
-                                    $this->_clearDirByKeepDays($fileBackupDir, $keepDays);
+                                    $this->_compressBackupFileDir($backupDirPath, date('Ymd', time() - 86400));
+                                    $this->_clearDirByKeepDays($fileBackupDir, $keepDays, $folderName);
                                 }
                             }
+                            $ssh = new Lib_Network_Ssh($fileConfig['host'], $fileConfig['port'], $fileConfig['user'], $fileConfig['pass']);
+                            // 先判断有没有安装sshpass工具，没有则自动安装
+                            $result = $ssh->checkCmd('sshpass');
+                            if (empty($result)) {
+                                $issueInfo = $ssh->syncCmd('cat /etc/issue');
+                                if (stripos($issueInfo, 'ubuntu') !== false) {
+                                    // Debian/Ubuntu 系统
+                                    $ssh->syncCmd("echo \"{$pass}\" | sudo -S apt-get install -y sshpass");
+                                } else {
+                                    // CentOS/RedHat 系统，注意这个时候只有root用户才能执行该命令
+                                    $ssh->syncCmd("yum install -y sshpass");
+                                }
+                            }
+                            $ssh->syncCmd("rsync -aurvz --delete -e 'sshpass -p {$pass} ssh -p {$port}' {$folderPath} {$user}@{$host}:{$fileBackupDir}");
                         } catch (\Exception $e) {
                             echo $e->getMessage().PHP_EOL;
                         }
@@ -100,31 +124,39 @@ class Controller_Backupper extends Controller_Base
      * 压缩备份的文件目录
      *
      * @param string $backupFileDirPath 文件目录绝对路径
+     * @param string $date              备份文件的日期
      *
      * @return void
      */
-    private function _compressBackupFileDir($backupFileDirPath)
+    private function _compressBackupFileDir($backupFileDirPath, $date)
     {
-        $date    = date('Ymd');
+        $currentDirPath = getcwd();
         $dirPath = dirname($backupFileDirPath);
         $dirName = basename($backupFileDirPath);
-        exec("tar -cjvf {$dirPath}/{$dirName}.{$date}.tar.bz2 {$backupFileDirPath}");
+        chdir($dirPath);
+        exec("tar -cjvf {$dirPath}/{$dirName}.{$date}.tar.bz2 {$dirName}");
+        chdir($currentDirPath);
     }
 
     /**
      * 按照给定天数清除超过保存期限的备份文件。
      *
-     * @param string  $dirPath  备份目录绝对路径
-     * @param integer $keepDays 保存天数
+     * @param string  $dirPath        备份目录绝对路径
+     * @param integer $keepDays       保存天数
+     * @param string  $fileNamePrefix 文件名前缀
      *
      * @return void
      */
-    private function _clearDirByKeepDays($dirPath, $keepDays)
+    private function _clearDirByKeepDays($dirPath, $keepDays, $fileNamePrefix = null)
     {
         $files = array_diff(scandir($dirPath), array('..', '.'));
         // 只计算压缩文件的数量
         foreach ($files as $k => $file) {
-            if (!preg_match('/.+\.bz2/', $file)) {
+            if (!empty($fileNamePrefix)) {
+                if (!preg_match("/^{$fileNamePrefix}.+\.bz2/", $file)) {
+                    unset($files[$k]);
+                }
+            } elseif (!preg_match('/.+\.bz2/', $file)) {
                 unset($files[$k]);
             }
         }
